+++ templates/_github/workflows/test-and-release.yml.ts
@@ -11,12 +11,20 @@ const templateFunction: TemplateFunction = answers => {
 		...(useTypeScript ? ["build/"] : []),
 		...(useReact ? ["admin/build/"] : []),
 	];
+	const useReleaseScript = answers.releaseScript === "yes";
+	const isGitHub = answers.target === "github";
 
 	const latestNodeVersion = "14.x";
 	const adapterTestVersions = ["12.x", "14.x", "16.x"];
 	const adapterTestOS = ["ubuntu-latest", "windows-latest", "macos-latest"];
 
 	const adapterName = answers.adapterName;
+
+	const deploy = useReleaseScript && isGitHub;
+	const escapeDeploy = 
+		deploy ?
+			(input: string) => input :
+			(input: string) => input.replace(/^/gm, '#');
 
 	const template = `
 name: Test and Release
@@ -69,44 +77,46 @@ ${isAdapter ? (
           # Uncomment the following line if your adapter cannot be installed using 'npm ci'
           # install-command: 'npm install'${needsBuild ? (`
           build: true`) : ""}
-`) : ""}
+`) : ""}${deploy ? "" : (`
 # TODO: To enable automatic npm releases, create a token on npmjs.org 
 # Enter this token as a GitHub secret (with name NPM_TOKEN) in the repository options
 # Then uncomment the following block:
+`)}
+${escapeDeploy(
+`  # Deploys the final package to NPM
+  deploy:
+    needs: [check-and-lint${isAdapter ? ", adapter-tests" : ""}]
 
-#  # Deploys the final package to NPM
-#  deploy:
-#    needs: [check-and-lint${isAdapter ? ", adapter-tests" : ""}]
-#
-#    # Trigger this step only when a commit on any branch is tagged with a version number
-#    if: |
-#      contains(github.event.head_commit.message, '[skip ci]') == false &&
-#      github.event_name == 'push' &&
-#      startsWith(github.ref, 'refs/tags/v')
-#
-#    runs-on: ubuntu-latest
-#
-#    steps:
-#      - uses: ioBroker/testing-action-deploy@v1
-#        with:
-#          node-version: '${latestNodeVersion}'
-#          # Uncomment the following line if your adapter cannot be installed using 'npm ci'
-#          # install-command: 'npm install'${needsBuild ? (`
-#          build: true`) : ""}
-#          npm-token: \${{ secrets.NPM_TOKEN }}
-#          github-token: \${{ secrets.GITHUB_TOKEN }}
-#
-#          # When using Sentry for error reporting, Sentry can be informed about new releases
-#          # To enable create a API-Token in Sentry (User settings, API keys)
-#          # Enter this token as a GitHub secret (with name SENTRY_AUTH_TOKEN) in the repository options
-#          # Then uncomment and customize the following block:
-#          sentry-token: \${{ secrets.SENTRY_AUTH_TOKEN }}
-#          sentry-project: "iobroker-${adapterName}"
-#          sentry-version-prefix: "iobroker.${adapterName}"${needsBuild ? (`
-#          sentry-sourcemap-paths: "${sourcemapPaths.join(" ")}"`) : ""}
-#          # If your sentry project is linked to a GitHub repository, you can enable the following option
-#          # sentry-github-integration: true
-`;	return template.trimLeft();
+    # Trigger this step only when a commit on any branch is tagged with a version number
+    if: |
+      contains(github.event.head_commit.message, '[skip ci]') == false &&
+      github.event_name == 'push' &&
+      startsWith(github.ref, 'refs/tags/v')
+
+    runs-on: ubuntu-latest
+
+    steps:
+      - uses: ioBroker/testing-action-deploy@v1
+        with:
+          node-version: '${latestNodeVersion}'
+          # Uncomment the following line if your adapter cannot be installed using 'npm ci'
+          # install-command: 'npm install'${needsBuild ? (`
+          build: true`) : ""}
+          npm-token: \${{ secrets.NPM_TOKEN }}
+          github-token: \${{ secrets.GITHUB_TOKEN }}
+
+          # When using Sentry for error reporting, Sentry can be informed about new releases
+          # To enable create a API-Token in Sentry (User settings, API keys)
+          # Enter this token as a GitHub secret (with name SENTRY_AUTH_TOKEN) in the repository options
+          # Then uncomment and customize the following block:
+          sentry-token: \${{ secrets.SENTRY_AUTH_TOKEN }}
+          sentry-project: "iobroker-${adapterName}"
+          sentry-version-prefix: "iobroker.${adapterName}"${needsBuild ? (`
+          sentry-sourcemap-paths: "${sourcemapPaths.join(" ")}"`) : ""}
+          # If your sentry project is linked to a GitHub repository, you can enable the following option
+          # sentry-github-integration: true`)}
+`;
+	return template.trimLeft();
 };
 
 templateFunction.customPath = ".github/workflows/test-and-release.yml";