+++ CHANGELOG.md
@@ -17,6 +17,8 @@
 * (AlCalzone) Remove `.npmignore` blacklist, use `files` whitelist in `package.json` instead (#816) · [Migration guide](docs/updates/20210906_files_whitelist.md)
 * (UncleSamSwiss) Updated alt texts of ioBroker badges in readme (#722)
 * (AlCalzone) Shared testing workflows (#822) · [Migration guide](docs/updates/20210913_shared_testing_workflows.md)
+* (UncleSamSwiss) Readme for adapter created using portal reflects the fact that the repository already exists (#788)
+* (UncleSamSwiss) The default branch can now be chosen, default is "main", but it is possible to change to "master" (#631)
 
 ## 1.34.1 (2021-07-07)
 * (UncleSamSwiss) Fix missing `common.main` in `io-package.json` (#778)